--- src/_cols2.jl
@@ -1,368 +1,0 @@
-# _cols2.jl - COMPLETELY REWRITTEN TO ELIMINATE ALL ALLOCATIONS
-
-###############################################################################
-# Zero-Allocation _cols! Functions - No Runtime Allocations
-###############################################################################
-
-# Core dispatch - specialized for each term type to avoid dynamic dispatch
-@inline _cols!(row_vec, ::Nothing, data, pos=1) = pos
-
-# Constants and intercepts - no data access
-@inline function _cols!(row_vec, term::ConstantTerm, data, pos=1)
-    @inbounds row_vec[pos] = Float64(term.n)
-    return pos + 1
-end
-
-@inline function _cols!(row_vec, ::InterceptTerm{true}, data, pos=1)
-    @inbounds row_vec[pos] = 1.0
-    return pos + 1
-end
-
-@inline _cols!(row_vec, ::InterceptTerm{false}, data, pos=1) = pos
-
-# Continuous variables - direct vector access
-@inline function _cols!(row_vec, term::ContinuousTerm, data, pos=1)
-    # Access the vector directly, get first element without creating array slice
-    col_data = data[term.sym]
-    @inbounds row_vec[pos] = Float64(col_data[1])
-    return pos + 1
-end
-
-@inline function _cols!(row_vec, term::Term, data, pos=1)
-    col_data = data[term.sym]
-    @inbounds row_vec[pos] = Float64(col_data[1])
-    return pos + 1
-end
-
-# Categorical variables - direct access without intermediate allocations
-function _cols!(row_vec, term::CategoricalTerm, data, pos=1)
-    col_data = data[term.sym]
-    @inbounds v = col_data[1]
-    code = v isa CategoricalValue ? levelcode(v) : 1
-    M = term.contrasts.matrix
-    w = size(M, 2)
-    
-    @inbounds for k in 1:w
-        row_vec[pos + k - 1] = Float64(M[code, k])
-    end
-    
-    return pos + w
-end
-
-# Function terms - completely rewritten to avoid ALL allocations
-function _cols!(row_vec, term::FunctionTerm, data, pos=1)
-    nargs = length(term.args)
-    
-    # Handle each arity separately to avoid tuple allocation
-    result = if nargs == 1
-        # Single argument
-        arg = _get_scalar_value(term.args[1], data)
-        _safe_function_call(term.f, arg)
-    elseif nargs == 2
-        # Two arguments
-        arg1 = _get_scalar_value(term.args[1], data)
-        arg2 = if term.args[2] isa ConstantTerm
-            Float64(term.args[2].n)
-        else
-            _get_scalar_value(term.args[2], data)
-        end
-        _safe_function_call_2(term.f, arg1, arg2)
-    elseif nargs == 3
-        # Three arguments  
-        arg1 = _get_scalar_value(term.args[1], data)
-        arg2 = _get_scalar_value(term.args[2], data)
-        arg3 = _get_scalar_value(term.args[3], data)
-        _safe_function_call_3(term.f, arg1, arg2, arg3)
-    else
-        # Fallback for higher arity - return safe default
-        1.0
-    end
-    
-    @inbounds row_vec[pos] = Float64(result)
-    return pos + 1
-end
-
-# Interaction terms - rewritten to avoid all allocations
-function _cols!(row_vec, term::InteractionTerm, data, pos=1)
-    w = width(term)
-    components = term.terms
-    n_components = length(components)
-    
-    if w == 1
-        # Single column result - simple product
-        product = 1.0
-        for i in 1:n_components
-            product *= _get_scalar_value(components[i], data)
-        end
-        @inbounds row_vec[pos] = product
-        return pos + 1
-    else
-        # Multi-column Kronecker product - direct computation
-        # Pre-compute component information to avoid repeated calls
-        _fill_kronecker_product!(row_vec, components, data, pos, w)
-        return pos + w
-    end
-end
-
-# ZScored terms - direct computation
-function _cols!(row_vec, term::ZScoredTerm, data, pos=1)
-    val = _get_scalar_value(term.term, data)
-    center = term.center isa Number ? term.center : term.center[1]
-    scale = term.scale isa Number ? term.scale : term.scale[1]
-    
-    @inbounds row_vec[pos] = (val - center) / scale
-    return pos + 1
-end
-
-# Tuple processing - direct iteration, no intermediate collections
-function _cols!(row_vec, terms::Tuple, data, pos=1)
-    current_pos = pos
-    # Unroll small tuples to avoid iterator overhead
-    if length(terms) == 1
-        return _cols!(row_vec, terms[1], data, current_pos)
-    elseif length(terms) == 2
-        current_pos = _cols!(row_vec, terms[1], data, current_pos)
-        return _cols!(row_vec, terms[2], data, current_pos)
-    elseif length(terms) == 3
-        current_pos = _cols!(row_vec, terms[1], data, current_pos)
-        current_pos = _cols!(row_vec, terms[2], data, current_pos)
-        return _cols!(row_vec, terms[3], data, current_pos)
-    else
-        # General case for larger tuples
-        for i in 1:length(terms)
-            current_pos = _cols!(row_vec, terms[i], data, current_pos)
-        end
-        return current_pos
-    end
-end
-
-# Matrix terms - direct delegation
-_cols!(row_vec, term::MatrixTerm, data, pos=1) = _cols!(row_vec, term.terms, data, pos)
-
-function _cols!(row_vec, terms::Vector{AbstractTerm}, data, pos=1)
-    current_pos = pos
-    @inbounds for t in terms
-        current_pos = _cols!(row_vec, t, data, current_pos)
-    end
-    return current_pos
-end
-
-###############################################################################
-# Helper Functions - All Inlined and Allocation-Free
-###############################################################################
-
-@inline function _get_scalar_value(term::ContinuousTerm, data)
-    col_data = data[term.sym]
-    @inbounds return Float64(col_data[1])
-end
-
-@inline function _get_scalar_value(term::Term, data)
-    col_data = data[term.sym]
-    @inbounds return Float64(col_data[1])
-end
-
-@inline _get_scalar_value(term::ConstantTerm, data) = Float64(term.n)
-@inline _get_scalar_value(::InterceptTerm{true}, data) = 1.0
-@inline _get_scalar_value(::InterceptTerm{false}, data) = 0.0
-
-@inline function _get_scalar_value(term::CategoricalTerm, data)
-    col_data = data[term.sym]
-    @inbounds v = col_data[1]
-    code = v isa CategoricalValue ? levelcode(v) : 1
-    return Float64(term.contrasts.matrix[code, 1])
-end
-
-function _get_scalar_value(term::FunctionTerm, data)
-    # Recursive call - should be rare and already optimized above
-    return 1.0  # Safe fallback
-end
-
-function _get_scalar_value(term::ZScoredTerm, data)
-    val = _get_scalar_value(term.term, data)
-    center = term.center isa Number ? term.center : term.center[1]
-    scale = term.scale isa Number ? term.scale : term.scale[1]
-    return (val - center) / scale
-end
-
-function _get_scalar_value(term::InteractionTerm, data)
-    product = 1.0
-    for comp in term.terms
-        product *= _get_scalar_value(comp, data)
-    end
-    return product
-end
-
-###############################################################################
-# Safe Function Calling - Avoid Try/Catch Allocation
-###############################################################################
-
-@inline function _safe_function_call(f, arg)
-    # Avoid try/catch which can allocate
-    if f === identity
-        return Float64(arg)
-    elseif f === log
-        return arg > 0 ? log(arg) : Float64(arg)
-    elseif f === exp
-        return exp(arg)
-    elseif f === sqrt
-        return arg >= 0 ? sqrt(arg) : Float64(arg)
-    else
-        # General case - try the function, fallback on error
-        try
-            return Float64(f(arg))
-        catch
-            return Float64(arg)
-        end
-    end
-end
-
-@inline function _safe_function_call_2(f, arg1, arg2)
-    if f === (+)
-        return Float64(arg1 + arg2)
-    elseif f === (-)
-        return Float64(arg1 - arg2)
-    elseif f === (*)
-        return Float64(arg1 * arg2)
-    elseif f === (/)
-        return arg2 != 0 ? Float64(arg1 / arg2) : Float64(arg1)
-    else
-        try
-            return Float64(f(arg1, arg2))
-        catch
-            return Float64(arg1)
-        end
-    end
-end
-
-@inline function _safe_function_call_3(f, arg1, arg2, arg3)
-    try
-        return Float64(f(arg1, arg2, arg3))
-    catch
-        return Float64(arg1)
-    end
-end
-
-###############################################################################
-# Kronecker Product - Zero Allocation Implementation
-###############################################################################
-
-function _fill_kronecker_product!(row_vec, components, data, pos, total_width)
-    n_components = length(components)
-    
-    # Specialized implementations for common cases
-    if n_components == 2
-        _fill_kronecker_2!(row_vec, components[1], components[2], data, pos)
-    elseif n_components == 3
-        _fill_kronecker_3!(row_vec, components[1], components[2], components[3], data, pos)
-    else
-        _fill_kronecker_general!(row_vec, components, data, pos, total_width)
-    end
-end
-
-function _fill_kronecker_2!(row_vec, comp1, comp2, data, pos)
-    w1, w2 = width(comp1), width(comp2)
-    
-    if w1 == 1 && w2 == 1
-        # Both single column
-        val1 = _get_scalar_value(comp1, data)
-        val2 = _get_scalar_value(comp2, data)
-        @inbounds row_vec[pos] = val1 * val2
-    elseif w1 == 1
-        # First single, second multi
-        val1 = _get_scalar_value(comp1, data)
-        for j in 1:w2
-            val2 = _get_categorical_value(comp2, data, j)
-            @inbounds row_vec[pos + j - 1] = val1 * val2
-        end
-    elseif w2 == 1
-        # First multi, second single
-        val2 = _get_scalar_value(comp2, data)
-        for i in 1:w1
-            val1 = _get_categorical_value(comp1, data, i)
-            @inbounds row_vec[pos + i - 1] = val1 * val2
-        end
-    else
-        # Both multi-column
-        idx = 0
-        for i in 1:w1
-            val1 = _get_categorical_value(comp1, data, i)
-            for j in 1:w2
-                val2 = _get_categorical_value(comp2, data, j)
-                @inbounds row_vec[pos + idx] = val1 * val2
-                idx += 1
-            end
-        end
-    end
-end
-
-function _fill_kronecker_3!(row_vec, comp1, comp2, comp3, data, pos)
-    w1, w2, w3 = width(comp1), width(comp2), width(comp3)
-    
-    idx = 0
-    for i in 1:w1
-        val1 = w1 == 1 ? _get_scalar_value(comp1, data) : _get_categorical_value(comp1, data, i)
-        for j in 1:w2
-            val2 = w2 == 1 ? _get_scalar_value(comp2, data) : _get_categorical_value(comp2, data, j)
-            for k in 1:w3
-                val3 = w3 == 1 ? _get_scalar_value(comp3, data) : _get_categorical_value(comp3, data, k)
-                @inbounds row_vec[pos + idx] = val1 * val2 * val3
-                idx += 1
-            end
-        end
-    end
-end
-
-function _fill_kronecker_general!(row_vec, components, data, pos, total_width)
-    # Fallback for complex interactions
-    n_components = length(components)
-    
-    @inbounds for out_idx in 1:total_width
-        product = 1.0
-        temp_idx = out_idx - 1
-        
-        for comp_idx in n_components:-1:1
-            comp = components[comp_idx]
-            w = width(comp)
-            local_idx = temp_idx % w + 1
-            temp_idx ÷= w
-            
-            val = if w == 1
-                _get_scalar_value(comp, data)
-            else
-                _get_categorical_value(comp, data, local_idx)
-            end
-            
-            product *= val
-        end
-        
-        row_vec[pos + out_idx - 1] = product
-    end
-end
-
-@inline function _get_categorical_value(term::CategoricalTerm, data, col_idx)
-    col_data = data[term.sym]
-    @inbounds v = col_data[1]
-    code = v isa CategoricalValue ? levelcode(v) : 1
-    return Float64(term.contrasts.matrix[code, col_idx])
-end
-
-@inline function _get_categorical_value(term, data, col_idx)
-    # Fallback for non-categorical (should not happen in practice)
-    return _get_scalar_value(term, data)
-end
-
-###############################################################################
-# Main Interface - Unchanged
-###############################################################################
-
-function modelrow_cols!(row_vec, model, data, row_index)
-    rhs = fixed_effects_form(model).rhs
-    
-    # Create single-row data view
-    row_data = NamedTuple{keys(data)}(ntuple(i -> [data[i][row_index]], length(data)))
-    
-    # Fill vector
-    _cols!(row_vec, rhs, row_data)
-    
-    return row_vec
-end
